fix(typings): theme context type issues
