+++ src/core/themes/themeContext.ts
@@ -1,12 +1,14 @@
 import React from 'react';
-import { ThemeKey } from '@src/core/themes/index';
+import { ThemeKey } from './themes';
 
-export interface ContextType {
+export interface ThemeContextType {
   currentTheme: ThemeKey;
-  toggleTheme: (theme: string) => void;
+  toggleTheme: (theme: ThemeKey) => void;
 }
 
-export const ThemeContext: React.Context<ContextType> = React.createContext({
+const initialValue: ThemeContextType = {
   currentTheme: 'Eva Light',
-  toggleTheme: (theme: string) => {},
-});
+  toggleTheme: (theme: ThemeKey) => {},
+};
+
+export const ThemeContext: React.Context<ThemeContextType> = React.createContext(initialValue);
